requirements.txt
amqp==5.1.0
asgiref==3.4.1
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.6
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0.1
docutils==0.17.1
ffmpeg==1.4
ffmpeg-python==0.2.0
flake8==4.0.1
future==0.18.2
idna==2.10
isort==5.10.1
kombu==5.2.4
loguru==0.5.3
mccabe==0.6.1
pexpect==4.8.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
Pygments==2.9.0
pyTelegramBotAPI==4.1.1
python-windscribe==2020.9.16
pytz==2022.1
qbittorrent-api==2021.12.26
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
srt==3.4.1
typing-extensions==4.2.0
tzdata==2021.5
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
win32-setctime==1.0.4
Redis
djangorestframework
django_celery_beat